refactor: remove storage of player and socket.id
Player i.e. player X or player O is already stored in the cookie
so no need to store it in the array.
Storage of socket.id is not required as the ID changes on reload
of the page and so is not a reliable means to track players.
This was why the move to cookies was decided on prior commits