feat:'X' appears to both players on square click
When a square is selected, the square index and the player cookie
is sent to the server. The server will store the player in the room
game data array. Following this the square number, room and the player
is returned to the client, which will check for players that contain
the specific room number and display 'X' in the correct square

Note in this case only 'X' has been defined as a player. Future commits
will create player 'O' which will be randomly assigned